# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: vrp-0 API
  version: 1.0.4-alpha-SNAPSHOT
...
paths:
...
  /matrix:
    get:
      tags:
      - 节点信息查询配置
      summary: 获取进程内 AMap 路由缓存矩阵
      description: 只访问 AmapTransitCalculator 的进程内缓存，不读取当前场景或求解任务中的矩阵。仅 MAP_PROVIDER=AMAP
        时可用。
      responses:
        "200":
          description: 进程内 AMap 路由缓存矩阵
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransitMatrix'
        "412":
          description: 当前地图提供方不是 AMap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
    post:
      tags:
      - 节点信息查询配置
      summary: 替换进程内 AMap 路由缓存矩阵
      description: 只更新 AmapTransitCalculator 的进程内缓存，不写入当前场景或求解任务中的矩阵。仅 MAP_PROVIDER=AMAP
        时可用。
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransitMatrix'
      responses:
        "200":
          description: 执行结果信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Msg'
        "412":
          description: 当前地图提供方不是 AMap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
  /mcp/doc:
    get:
      tags:
      - 节点信息查询配置
      summary: 打开 MCP 参考文档
      responses:
        "200":
          description: MCP Markdown 文档
        "404":
          description: 文档不存在
...
  /pois:
    get:
      tags:
      - 地址查询生成
      summary: 按关键词查询 POI
      parameters:
      - name: city
        in: query
        description: 城市/adcode
        schema:
          type: string
      - name: keywords
        in: query
        description: 关键词
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: 累计查询页数
        schema:
          format: int64
          default: 1
          minimum: 1
          type: integer
      - name: types
        in: query
        description: 类型
        schema:
          type: string
      responses:
        "200":
          description: 真实地址列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/POI'
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "412":
          description: 当前地图提供方不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "500":
          description: 请求当前地图提供方接口异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
  /pois/geocode:
    get:
      tags:
      - 地址查询生成
      summary: 按地址文本执行地理编码
      parameters:
      - name: city
        in: query
        description: 城市/adcode
        schema:
          type: string
      - name: keywords
        in: query
        description: 关键词
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 地理编码候选结果
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/POI'
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "412":
          description: 当前地图提供方不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "500":
          description: 请求当前地图提供方接口异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
  /pois/regeocode:
    get:
      tags:
      - 地址查询生成
      summary: 按经纬度执行逆地理编码
      parameters:
      - name: location
        in: query
        description: "经纬度，格式：lng,lat"
        required: true
        schema:
          type: string
      responses:
        "200":
          description: 逆地理编码结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/POI'
        "400":
          description: 参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "412":
          description: 当前地图提供方不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "500":
          description: 请求当前地图提供方接口异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
  /quota:
    get:
      tags:
      - 节点信息查询配置
      summary: 查询API私钥及限流策略（前端配置页口径）
      responses:
        "200":
          description: API私钥及限流策略
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conf'
    put:
      tags:
      - 节点信息查询配置
      summary: 更新API私钥及限流策略
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Conf'
      responses:
        "200":
          description: 已保存的API私钥及限流策略
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conf'
...
components:
  schemas:
...
    Conf:
      type: object
      properties:
        key:
          type: string
        qps:
          format: int32
          type: integer
        quota:
          format: int64
          type: integer
        interval:
          $ref: '#/components/schemas/Duration'
        wait_timeout:
          $ref: '#/components/schemas/Duration'
        geocode_provider:
          $ref: '#/components/schemas/GeocodeProvider'
        address_resolver_url:
          type: string
        address_resolver_fallback_to_amap:
          type: boolean
...
    Duration:
      format: duration
      type: string
      example: P1D
    ErrorInfo:
      description: 统一 REST 错误响应。网页客户端使用 error_code 和 error_params 本地化展示，message 仅保留给兼容客户端和诊断。
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        error_code:
          description: 稳定、面向机器的错误码
          type: string
        error_params:
          description: 错误码插值所需的非敏感参数
          type: object
          additionalProperties:
            type: string
        message:
          description: 向后兼容的服务端诊断消息；网页不得直接展示
          type: string
...
    GeocodeProvider:
      enum:
      - AMAP
      - ADDR_RESOLVER
      type: string
...
    LOC:
      type: object
      properties:
        lat:
          format: double
          type: number
        lon:
          format: double
          type: number
...
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
...
    Msg:
      type: object
      properties:
        code:
          format: int32
          type: integer
        msg:
          type: string
    POI:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          description: 经纬度
          type: string
        loc:
          $ref: '#/components/schemas/LOC'
        address:
          description: 详细地址
          type: string
        pcode:
          description: 所属省份编码
          type: string
        type:
          description: 所属类型
          type: string
        photos:
          description: ' poi 图片相关信息'
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        gridcode:
          description: ' poi 的地理格 id'
          type: string
        typecode:
          description: 分类编码
          type: string
        citycode:
          description: 所属城市编码
          type: string
        adname:
          description: 所属区域名称
          type: string
        entr_location:
          description: 入口经纬度坐标
          type: string
        entr_loc:
          $ref: '#/components/schemas/LOC'
        adcode:
          description: 所属区域编码
          type: string
        pname:
          description: 所属省份名称
          type: string
        cityname:
          description: 所属城市名称
          type: string
    Photo:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
...
    Transit:
      type: object
      properties:
        distance:
          format: int64
          type: integer
        duration:
          format: int64
          type: integer
        create_time:
          $ref: '#/components/schemas/LocalDateTime'
    TransitMatrix:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              $ref: '#/components/schemas/Transit'
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...